Refactored FileBasedStateRepository and ReloadablePropertiesFile to use a new PropertySource to allow for other property-based repositories other than files.
